fix(api): bug-list #1 (TINYINT→SMALLINT) + #13 (rate limiter)
#1: bgpSessions.status/policy used DataTypes.TINYINT.UNSIGNED, which
Postgres doesn't support — a fresh sync({alter:true}) would fail with
`type "tinyint" does not exist`. Use SMALLINT. No effect on the running
prod DB (NODE_ENV=production skips sync); fixes fresh/CI/test deploys.

#13: rate limiter never applied its route limits — ROUTE_LIMITS keys were
"/agent" etc. but real paths are "/api/v1/agent", so path.startsWith never
matched and every route fell back to default (60/min). Fixed prefixes, and
exempt trusted internal clients (bot + agents authenticate with the agent
API key) so legitimate high-frequency core traffic (heartbeats, per-command
admin calls) is never 429'd. Public traffic (/auth login, /session user
JWTs, anonymous) still gets limited — which is the actual point.

+++ packages/api/src/db/models/bgpSessions.ts
@@ -73,7 +73,9 @@ export function initBgpSessionsModel(sequelize: Sequelize): BgpSessionsModel {
 				allowNull: false,
 			},
 			status: {
-				type: DataTypes.TINYINT.UNSIGNED,
+				// SMALLINT, not TINYINT: Postgres has no TINYINT, so a fresh
+				// sync({alter:true}) would fail with `type "tinyint" does not exist`.
+				type: DataTypes.SMALLINT.UNSIGNED,
 				allowNull: false,
 				defaultValue: PeeringStatus.PENDING_REVIEW,
 			},
@@ -83,7 +85,8 @@ export function initBgpSessionsModel(sequelize: Sequelize): BgpSessionsModel {
 				defaultValue: 1420,
 			},
 			policy: {
-				type: DataTypes.TINYINT.UNSIGNED,
+				// SMALLINT, not TINYINT (Postgres has no TINYINT) — see status above.
+				type: DataTypes.SMALLINT.UNSIGNED,
 				allowNull: false,
 				defaultValue: SessionPolicy.FULL,
 			},
